Move some utils to pulp_smash.selectors
Create module `pulp_smash.selectors`. This module contains "tools for selecting
and deselecting tests." Update the documentation, and update all references to
these functions from unit tests and functional tests. Test suite results do not
change:

    ============  =============================
    Pulp Version  Test Suite Results
    ============  =============================
    2.7           OK (skipped=14)
    dev (2.8)     FAILED (errors=2, skipped=13)
    ============  =============================
