+++ docs/api/pulp_smash.selectors.rst
@@ -1,0 +1,6 @@
+`pulp_smash.selectors`
+======================
+
+Location: :doc:`/index` → :doc:`/api` → :doc:`/api/pulp_smash.selectors`
+
+.. automodule:: pulp_smash.selectors
