+++ docs/api/tests.test_selectors.rst
@@ -1,0 +1,6 @@
+`tests.test_selectors`
+======================
+
+Location: :doc:`/index` → :doc:`/api` → :doc:`/api/tests.test_selectors`
+
+.. automodule:: tests.test_selectors
